requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
comm==0.1.3
ddddocr==1.4.7
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.1.1
executing==1.2.0
Flask==2.3.2
flatbuffers==23.5.9
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
humanfriendly==10.0
idna==3.4
ipykernel==6.23.1
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
line-bot-sdk==2.4.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.3
onnxruntime==1.14.1
opencv-python-headless==4.7.0.72
outcome==1.2.0
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
protobuf==4.23.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyimgur==0.6.0
PySocks==1.7.1
python-dateutil==2.8.2
pyzmq==25.0.2
requests==2.28.2
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.2
sympy==1.12
tornado==6.3.2
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
undetected-chromedriver==3.4.7
urllib3==1.26.15
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.4
wsproto==1.2.0
yarl==1.9.2
